Remove Plots from benchmarks and separate benchmark plotting
